+++ src/main/java/com/example/onehada/db/repository/AgentRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface AgentRepository extends JpaRepository<Agent, Integer> {
+public interface AgentRepository extends JpaRepository<Agent, Long> {
 	Optional<Agent> findByAgentEmailAndAgentPw(String email, String password);
 	List<Agent> findByAgentEmailContaining(String email);
 	List<Agent> findByAgentNameContaining(String name);
